новая структура, list Теперь icollection
