+++ SurveyPlatform.Infrastructure/Entities/PollResponse.cs
@@ -4,14 +4,16 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SurveyPlatform.Core.Entities
+namespace SurveyPlatform.DAL.Entities
 {
-    public class PollOption
+    public class PollResponse
     {
         public int Id { get; set; }
-        public string Content { get; set; }
         public int PollId { get; set; }
         public Poll Poll { get; set; }
-        public List<PollResponse> Responses { get; set; } // Добавлено
+        public int OptionId { get; set; }
+        public PollOption Option { get; set; }
+        public int UserId { get; set; }
+        public User User { get; set; }
     }
 }
