+++ SurveyPlatform.Infrastructure/Interfaces/IUserRepository.cs
@@ -1,11 +1,11 @@
-﻿using SurveyPlatform.Core.Entities;
+﻿using SurveyPlatform.DAL.Entities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SurveyPlatform.Core.Interfaces
+namespace SurveyPlatform.DAL.Interfaces
 {
     public interface IUserRepository
     {
