+++ SurveyPlatform.sln
@@ -5,9 +5,7 @@ VisualStudioVersion = 17.10.34928.147
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SurveyPlatform.API", "SurveyPlatform\SurveyPlatform.API.csproj", "{31416474-FE47-4227-98B2-90F111559C7E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SurveyPlatform.Core", "SurveyPlatform.Core\SurveyPlatform.Core.csproj", "{0DBC406F-592F-4E94-BE1C-A13EDB3FACB9}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SurveyPlatform.Infrastructure", "SurveyPlatform.Infrastructure\SurveyPlatform.Infrastructure.csproj", "{059B915D-2D27-4088-BA11-8043CAC3A9BB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SurveyPlatform.DAL", "SurveyPlatform.Infrastructure\SurveyPlatform.DAL.csproj", "{059B915D-2D27-4088-BA11-8043CAC3A9BB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -19,10 +17,6 @@ Global
 		{31416474-FE47-4227-98B2-90F111559C7E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{31416474-FE47-4227-98B2-90F111559C7E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{31416474-FE47-4227-98B2-90F111559C7E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0DBC406F-592F-4E94-BE1C-A13EDB3FACB9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0DBC406F-592F-4E94-BE1C-A13EDB3FACB9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0DBC406F-592F-4E94-BE1C-A13EDB3FACB9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0DBC406F-592F-4E94-BE1C-A13EDB3FACB9}.Release|Any CPU.Build.0 = Release|Any CPU
 		{059B915D-2D27-4088-BA11-8043CAC3A9BB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{059B915D-2D27-4088-BA11-8043CAC3A9BB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{059B915D-2D27-4088-BA11-8043CAC3A9BB}.Release|Any CPU.ActiveCfg = Release|Any CPU
